lista alunos quase finalizado
