+++ src/screens/ListaAlunos/components/FiltroAlunos/style.scss
@@ -1,4 +1,4 @@
-@import "../../style/variables.scss";
+@import "../../../../style/variables.scss";
 
 .list-filter {
   .top-collapse {
