+++ src/screens/ListaAlunos/components/TabelaResultado/helper.js
@@ -1,0 +1,8 @@
+export const getColor = status => {
+  switch (status) {
+    case "":
+      return "green";
+    default:
+      return "green";
+  }
+};
