+++ src/screens/ListaAlunos/components/TabelaResultado/style.scss
@@ -1,0 +1,63 @@
+@import "../../../../style/variables.scss";
+
+div.table-results {
+  div.title {
+    i {
+      padding-right: 10px;
+    }
+    color: $blackInput;
+    font-weight: bold;
+    font-size: 16px;
+  }
+  table.table-students {
+    width: 100%;
+    border-radius: 4px;
+    thead {
+      tr {
+        border-top-left-radius: 4px;
+        border-top-right-radius: 4px;
+        border: 1px solid $grayDisabled;
+        th {
+          font-weight: bold;
+          border-right: 1px solid $grayDisabled;
+          background-color: $grayCard;
+          color: $blackInput;
+          font-size: 14px;
+          padding: 10px 0;
+          padding-left: 10px;
+        }
+      }
+    }
+    tbody {
+      tr {
+        border: 1px solid $grayDisabled;
+        td {
+          border-right: 1px solid $grayDisabled;
+          color: $blackInput;
+          font-size: 14px;
+          padding: 10px;
+          padding-left: 10px;
+          &:not(:last-child) {
+            padding-top: 1.5em;
+          }
+          &.status {
+            font-size: 12px;
+            span {
+              padding: 5px;
+              border-radius: 4px;
+              background-color: $grayCard;
+              &.green {
+                color: $green;
+                border: 1px solid $green;
+              }
+            }
+          }
+        }
+        &:last-child {
+          border-bottom-left-radius: 4px;
+          border-bottom-right-radius: 4px;
+        }
+      }
+    }
+  }
+}
